Separate flow manager to separate files
