+++ src/Magenta.Workflow/IFlowSecurityContext{T}.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Magenta.Workflow.Managers.Flows;
+using Magenta.Workflow.Managers.Reports;
 
 namespace Magenta.Workflow
 {
@@ -22,7 +23,7 @@ namespace Magenta.Workflow
         #endregion
 
         #region Filter Access Cartable
-        Task<IQueryable<FlowCartableModel>> FilterAccessCartableAsync(IQueryable<FlowCartableModel> query);
+        Task<IQueryable<FlowInboxModel>> FilterAccessCartableAsync(IQueryable<FlowInboxModel> query);
         #endregion
     }
 }
