+++ src/Magenta.Workflow/Managers/Flows/FlowIdentityManager.cs
@@ -1,0 +1,9 @@
+﻿namespace Magenta.Workflow.Managers.Flows
+{
+    public partial class FlowManager
+    {
+
+
+
+    }
+}
